ngd-property-pipeline/mapper/requirements.txt
rdflib
telicent-ies-tool==1.0.2
telicent-lib==2.0.1
python-dotenv
setuptools
pytest
importlib
